requirements.txt
streamlit 
pandas 
plotly 
spacy 
google-generativeai 
openpyxl
python-dotenv      
transformers       

